+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 # Next
+## Added
+- Missing Django migration
 ## Changed
 - Allow Django 4.1+ (but not 5.0)
 - Add Python 3.11 in classifier